pull out inner static classes
